fix:delete space in byId and rename func in stac
